requirements.txt
asgiref==3.2.10
astroid==2.4.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==2.2.14
django-cors-headers==3.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-client==0.0.1
mysqlclient==2.0.1
numpy==1.19.0
pylint==2.5.2
pytz==2020.1
requests==2.24.0
scapy==2.4.4
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1